导出教材申报信息接口新增筛选条件：是否已申报教材

+++ src/main/java/xyz/kingsword/course/util/BookUtil.java
@@ -56,7 +56,7 @@ public class BookUtil {
                     stringBuilder.append(priceStr.charAt(i));
                 }
             }
-            book.setPrice(Float.parseFloat(stringBuilder.toString()));
+            book.setPrice(Double.parseDouble(stringBuilder.toString()));
         }
         return book;
     }
